+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Display UpdateOrderStatusModal on ApproveAndPay if the conditions are met. Refs UINV-577.
 * Add columns to invoice search results for vendor name and invoice number. Refs UINV-591.
 * Fix Tags filter query to avoid global OR. Refs UINV-603.
+* Fix invoice fund distr column in csv export. Refs UINV-604.
 
 ## [7.0.4](https://github.com/folio-org/ui-invoice/tree/v7.0.4) (2025-04-28)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v7.0.3...v7.0.4)